Update go.mod and go.sum to latest version from networkservicemesh/sdk-vpp@main PR link: https://github.com/networkservicemesh/sdk-vpp/pull/452
Commit: 7958db6
Author: Ed Warnicke
Date: 2021-11-22 09:09:20 -0600
Message:
  - Fix for mtu issues with kernelvethpair (#452)
This *probably* fixes:
https://github.com/networkservicemesh/sdk/issues/1148

The underlying issue was that the end of the vethpair to which
VPP was attaching with afpacket was not getting its MTU set correctly.
As a result, if an oversized packet was sent over it, it would be fragmented
by the kernel to a size that matches the mtu on the end of the veth pair
that was in the NSC network namespace.

The resulting packet would *still* be to large for the MTU of the end of the
vethpair attached to the VPP instance, and would be dropped there.

The second fragment, being smaller, would be smaller than the MTU of the
end of the vethpair to which VPP was attached with af-packet, and so get
through.

Signed-off-by: NSMBot <nsmbot@networkservicmesh.io>